+++ eq_cir_proxy_service/routers/instrument.py
@@ -7,9 +7,9 @@ from structlog import get_logger
 
 from eq_cir_proxy_service.exceptions import exception_messages
 from eq_cir_proxy_service.services.instrument import (
-    conversion,
-    metadata,
-    retrieval,
+    convert_instrument,
+    retrieve_instrument,
+    retrieve_instrument_metadata,
 )
 from eq_cir_proxy_service.services.validators.request import (
     validate_version,
@@ -36,11 +36,11 @@ async def get_instrument_by_uuid(
         validate_version(version)
         target_version = version
 
-        instrument = await retrieval.retrieve_instrument(instrument_id)
+        instrument = await retrieve_instrument.retrieve_instrument(instrument_id)
 
-        instrument_metadata = await metadata.retrieve_instrument_metadata(instrument_id)
+        instrument_metadata = await retrieve_instrument_metadata.retrieve_instrument_metadata(instrument_id)
 
-        return await conversion.convert_instrument(
+        return await convert_instrument.convert_instrument(
             instrument=instrument,
             instrument_metadata=instrument_metadata,
             target_version=target_version,
